Small fixes in format-xml.sh
- check if any *.xlsx exists in the current directory
- create `extract` directory with `-p` option, so no warning is shown if the directory already exists
- check if target directory exists before trying to remove it

+++ format-xml.sh
@@ -1,12 +1,13 @@
 #!/bin/bash
 
 # Helper script to extract test files to make them easier to visually inspect
+ls -q  -- *.xlsx > /dev/null || exit 1
 
-mkdir extract
+mkdir -p extract
 
 for x in *.xlsx ; do
 
-    rm -r extract/$x
+    test -d extract/$x && rm -r extract/$x
     mkdir extract/$x
     cd extract/$x
     unzip ../../$x
